listener: match rebalancer to listener IP family type (#16914)
When getting a rebalancer by address and a wild card match is being
used, the first match in the list is returned. However, if there are
listeners with addresses "0.0.0.0" and "::" then the first active
listener found will be used, irrespective of the IP family type. Change
the behavior to always return the listener of the same IP family type as
the rebalancer.

Fixes #16804

Co-authored-by: yingchun.cai@volunteers.acasi.info
Signed-off-by: Jacob Delgado <jacob.delgado@volunteers.acasi.info>