ext_proc: Implement STREAMED body processing mode (#17069)
Implement this mode, which allows an external processor to receive the body one chunk at a time and examine or modify the body while it is being processed.

Risk Level: Medium. The mode is only engaged if used. However an external processor must be coded carefully, particularly if it expects to stream both request and response bodies simultaneously, since each will proceed in its own ordering.

Testing: New unit and integration tests.

Docs Changes: Updated proto docs to reflect that the mode is now supported.

Release Notes: External processing servers may now use the STREAMED processing mode. In this mode, chunks of the body are forwarded to the external processing server when they arrive. Depending on how the upstream system is implemented, request body chunks may be delivered before or after the response headers, and request and response body chunks may be interleaved if the upstream system delivers them that way. An external processor should be carefully coded so that it does not assume that a particular ordering will be implemented.

Signed-off-by: Gregory Brail <gregbrail@google.com>