test: allocate X509 objects correctly (#17688)
As with any other C API, simply default-initializing an X509 will not
correctly set up the object. It will also break in the future. BoringSSL
is also looking to align with OpenSSL and make the X509 type opaque, to
avoid callers reaching into private fields and breaking things.

Instead, construct an X509 way with X509_new(). It's still an empty
object, which isn't really a certificate to verify, but this test
was already doing that and doesn't seem to care.

Signed-off-by: David Benjamin <davidben@google.com>